+++ client/components/index.js
@@ -7,9 +7,6 @@ export {default as Navbar} from './navbar/Navbar'
 export {default as UserHome} from './user/user-home'
 export {Login, Signup} from './auth/auth-form'
 export {default as AddressBook} from './sidePanel/addressBook/AddressBook'
-export {
-  default as CategoryFilter
-} from './sidePanel/categoryFilter/CategoryFilter'
 export {default as Categories} from './sidePanel/categoryFilter/Categories'
 export {default as Map} from './map/Map'
 export {default as Welcome} from './welcome/Welcome'
